drivers: plic: Fix parameter type of plic_op_raise_sgi()
The commit ec740b9fe95e ("core: interrupt_raise_sgi() updates") changes
the cpu_mask parameter to a uint32_t. Apply this change onto
plic_op_raise_sgi().

Acked-by: Jens Wiklander <jens.wiklander@linaro.org>

+++ core/drivers/plic.c
@@ -197,7 +197,7 @@ static void plic_op_raise_pi(struct itr_chip *chip, size_t it)
 }
 
 static void plic_op_raise_sgi(struct itr_chip *chip __unused,
-			      size_t it __unused, uint8_t cpu_mask __unused)
+			      size_t it __unused, uint32_t cpu_mask __unused)
 {
 }
